requirements.txt
aiofiles==24.1.0 ; python_version >= "3.12" and python_version < "3.14"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "3.14"
anthropic==0.64.0 ; python_version >= "3.12" and python_version < "3.14"
anyio==4.10.0 ; python_version >= "3.12" and python_version < "3.14"
audioop-lts==0.2.2 ; python_version == "3.13"
brotli==1.1.0 ; python_version >= "3.12" and python_version < "3.14"
certifi==2025.8.3 ; python_version >= "3.12" and python_version < "3.14"
charset-normalizer==3.4.3 ; python_version >= "3.12" and python_version < "3.14"
click==8.2.1 ; python_version >= "3.12" and python_version < "3.14" and sys_platform != "emscripten"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "3.14" and platform_system == "Windows"
distro==1.9.0 ; python_version >= "3.12" and python_version < "3.14"
fastapi==0.116.1 ; python_version >= "3.12" and python_version < "3.14"
ffmpy==0.6.1 ; python_version >= "3.12" and python_version < "3.14"
filelock==3.19.1 ; python_version >= "3.12" and python_version < "3.14"
fsspec==2025.7.0 ; python_version >= "3.12" and python_version < "3.14"
gradio-client==1.12.1 ; python_version >= "3.12" and python_version < "3.14"
gradio==5.44.0 ; python_version >= "3.12" and python_version < "3.14"
groovy==0.1.2 ; python_version >= "3.12" and python_version < "3.14"
h11==0.16.0 ; python_version >= "3.12" and python_version < "3.14"
hf-xet==1.1.9 ; python_version >= "3.12" and python_version < "3.14" and (platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "arm64" or platform_machine == "aarch64")
httpcore==1.0.9 ; python_version >= "3.12" and python_version < "3.14"
httpx==0.28.1 ; python_version >= "3.12" and python_version < "3.14"
huggingface-hub==0.34.4 ; python_version >= "3.12" and python_version < "3.14"
idna==3.10 ; python_version >= "3.12" and python_version < "3.14"
jinja2==3.1.6 ; python_version >= "3.12" and python_version < "3.14"
jiter==0.10.0 ; python_version >= "3.12" and python_version < "3.14"
markdown-it-py==4.0.0 ; python_version >= "3.12" and python_version < "3.14" and sys_platform != "emscripten"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "3.14"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "3.14" and sys_platform != "emscripten"
numpy==2.3.2 ; python_version >= "3.12" and python_version < "3.14"
orjson==3.11.3 ; python_version >= "3.12" and python_version < "3.14"
packaging==25.0 ; python_version >= "3.12" and python_version < "3.14"
pandas==2.3.2 ; python_version >= "3.12" and python_version < "3.14"
pillow==11.3.0 ; python_version >= "3.12" and python_version < "3.14"
pydantic-core==2.33.2 ; python_version >= "3.12" and python_version < "3.14"
pydantic==2.11.7 ; python_version >= "3.12" and python_version < "3.14"
pydub==0.25.1 ; python_version >= "3.12" and python_version < "3.14"
pygments==2.19.2 ; python_version >= "3.12" and python_version < "3.14" and sys_platform != "emscripten"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "3.14"
python-dotenv==1.1.1 ; python_version >= "3.12" and python_version < "3.14"
python-multipart==0.0.20 ; python_version >= "3.12" and python_version < "3.14"
pytz==2025.2 ; python_version >= "3.12" and python_version < "3.14"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "3.14"
requests==2.32.5 ; python_version >= "3.12" and python_version < "3.14"
rich==14.1.0 ; python_version >= "3.12" and python_version < "3.14" and sys_platform != "emscripten"
ruff==0.12.10 ; python_version >= "3.12" and python_version < "3.14" and sys_platform != "emscripten"
safehttpx==0.1.6 ; python_version >= "3.12" and python_version < "3.14"
semantic-version==2.10.0 ; python_version >= "3.12" and python_version < "3.14"
shellingham==1.5.4 ; python_version >= "3.12" and python_version < "3.14" and sys_platform != "emscripten"
six==1.17.0 ; python_version >= "3.12" and python_version < "3.14"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "3.14"
starlette==0.47.3 ; python_version >= "3.12" and python_version < "3.14"
tomlkit==0.13.3 ; python_version >= "3.12" and python_version < "3.14"
tqdm==4.67.1 ; python_version >= "3.12" and python_version < "3.14"
typer==0.16.1 ; python_version >= "3.12" and python_version < "3.14" and sys_platform != "emscripten"
typing-extensions==4.15.0 ; python_version >= "3.12" and python_version < "3.14"
typing-inspection==0.4.1 ; python_version >= "3.12" and python_version < "3.14"
tzdata==2025.2 ; python_version >= "3.12" and python_version < "3.14"
urllib3==2.5.0 ; python_version >= "3.12" and python_version < "3.14"
uvicorn==0.35.0 ; python_version >= "3.12" and python_version < "3.14" and sys_platform != "emscripten"
websockets==15.0.1 ; python_version >= "3.12" and python_version < "3.14"
yt-dlp==2025.8.22 ; python_version >= "3.12" and python_version < "3.14"
